NN-2115: Rename PayOptions and PayOtherForm to AttendOptions and AttendOtherForm
