+++ src/ResultsActivity/elements/AttendanceOptions/AttendanceOptions.styles.js
