+++ src/ResultsActivity/elements/AttendanceOptions/index.js
@@ -1,0 +1,3 @@
+import AttendanceOptions from './AttendanceOptions'
+
+export default AttendanceOptions
