+++ src/ResultsActivity/elements/AttendanceOtherForm/index.js
@@ -1,0 +1,3 @@
+import AttendanceOtherForm from './AttendanceOtherForm'
+
+export default AttendanceOtherForm
