--- src/ResultsActivity/elements/PayOptions/index.js
@@ -1,3 +1,0 @@
-import PayOptions from './PayOptions'
-
-export default PayOptions
