--- src/ResultsActivity/elements/PayOtherForm/index.js
@@ -1,3 +1,0 @@
-import PayOtherForm from './PayOtherForm'
-
-export default PayOtherForm
